Redirect Batch
This is a php file that scrapes your database for a Name and URL,
constructs an index.htm file with the proper re-direct code... Then, it
creates a folder (this uses the NAME field in the database) where it
saves the index.htm file in.

It's a pretty simple script, nothing major and even easier to modify.